+++ sparkmagic/sparkmagic/kernels/kernelmagics.py
@@ -11,7 +11,7 @@ from IPython.core.magic import needs_local_scope, cell_magic
 from IPython.core.magic_arguments import argument, magic_arguments
 from hdijupyterutils.utils import generate_uuid
 
-from sparkmagic.utils.configuration import SparkMagicConfiguration
+import sparkmagic.utils.configuration as conf
 from sparkmagic.utils.utils import get_livy_kind, parse_argstring_or_throw
 from sparkmagic.utils.sparkevents import SparkEvents
 from sparkmagic.utils.constants import LANGS_SUPPORTED
@@ -47,8 +47,6 @@ class KernelMagics(SparkMagicBase):
         # You must call the parent constructor
         super(KernelMagics, self).__init__(shell, data)
         
-        self.conf = SparkMagicConfiguration()
-
         self.session_name = u"session_name"
         self.session_started = False
 
@@ -146,7 +144,7 @@ class KernelMagics(SparkMagicBase):
         else:
             current_session_id = None
 
-        self.ipython_display.html(u"Current session configs: <tt>{}</tt><br>".format(self.conf.get_session_properties(self.language)))
+        self.ipython_display.html(u"Current session configs: <tt>{}</tt><br>".format(conf.get_session_properties(self.language)))
 
         info_sessions = self.spark_controller.get_all_sessions_endpoint(self.endpoint)
         self._print_endpoint_info(info_sessions, current_session_id)
@@ -287,14 +285,14 @@ class KernelMagics(SparkMagicBase):
 
         if not self.session_started:
             skip = False
-            properties = self.conf.get_session_properties(self.language)
+            properties = conf.get_session_properties(self.language)
             self.session_started = True
 
             try:
                 self.spark_controller.add_session(self.session_name, self.endpoint, skip, properties)
             except Exception as e:
                 self.fatal_error = True
-                self.fatal_error_message = self.conf.fatal_error_suggestion().format(e)
+                self.fatal_error_message = conf.fatal_error_suggestion().format(e)
                 self.logger.error(u"Error creating session: {}".format(e))
                 self.ipython_display.send_error(self.fatal_error_message)
                 return False
@@ -350,7 +348,7 @@ class KernelMagics(SparkMagicBase):
 
     @staticmethod
     def _override_session_settings(settings):
-        self.conf.override(self.conf.session_configs.__name__, settings)
+        conf.override(conf.session_configs.__name__, settings)
 
     @staticmethod
     def _generate_uuid():
