+++ sparkmagic/sparkmagic/kernels/wrapperkernel/sparkkernelbase.py
@@ -5,8 +5,8 @@ from ipykernel.ipkernel import IPythonKernel
 from hdijupyterutils.ipythondisplay import IpythonDisplay
 from hdijupyterutils.log import Log
 
+import sparkmagic.utils.configuration as conf
 from sparkmagic.utils.constants import MAGICS_LOGGER_NAME
-from sparkmagic.utils.configuration import SparkMagicConfiguration
 from sparkmagic.livyclientlib.exceptions import wrap_unexpected_exceptions
 from sparkmagic.kernels.wrapperkernel.usercodeparser import UserCodeParser
 
@@ -26,9 +26,7 @@ class SparkKernelBase(IPythonKernel):
 
         super(SparkKernelBase, self).__init__(**kwargs)
 
-        self.conf = SparkMagicConfiguration()
-
-        self.logger = Log(MAGICS_LOGGER_NAME, self.conf.logging_config(), u"_jupyter_kernel".format(self.session_language))
+        self.logger = Log(MAGICS_LOGGER_NAME, conf.logging_config(), u"_jupyter_kernel".format(self.session_language))
         self._fatal_error = None
         self.ipython_display = IpythonDisplay()
 
@@ -43,7 +41,7 @@ class SparkKernelBase(IPythonKernel):
         if not kwargs.get("testing", False):
             self._load_magics_extension()
             self._change_language()
-            if self.conf.use_auto_viz():
+            if conf.use_auto_viz():
                 self._register_auto_viz()
 
     def do_execute(self, code, silent, store_history=True, user_expressions=None, allow_stdin=False):
@@ -132,7 +130,7 @@ ip.display_formatter.ipython_display_formatter.for_type_by_name('pandas.core.fra
 
     def _repeat_fatal_error(self):
         """Throws an error that has already been queued."""
-        error = self.conf.fatal_error_suggestion().format(self._fatal_error)
+        error = conf.fatal_error_suggestion().format(self._fatal_error)
         self.logger.error(error)
         self.ipython_display.send_error(error)
         return self._complete_cell()
