+++ sparkmagic/sparkmagic/livyclientlib/reliablehttpclient.py
@@ -5,9 +5,8 @@ from time import sleep
 import requests
 from hdijupyterutils.log import Log
 
-from sparkmagic.utils.configuration import SparkMagicConfiguration
+import sparkmagic.utils.configuration as conf
 from sparkmagic.utils.constants import MAGICS_LOGGER_NAME
-from sparkmagic.utils.configuration import SparkMagicConfiguration
 from sparkmagic.livyclientlib.exceptions import HttpClientException
 
 
@@ -18,10 +17,9 @@ class ReliableHttpClient(object):
         self._endpoint = endpoint
         self._headers = headers
         self._retry_policy = retry_policy
-        self.logger = Log(MAGICS_LOGGER_NAME, SparkMagicConfiguration().logging_config(), u"ReliableHttpClient")
-        self.conf = SparkMagicConfiguration()
+        self.logger = Log(MAGICS_LOGGER_NAME, conf.logging_config(), u"ReliableHttpClient")
 
-        self.verify_ssl = not self.conf.ignore_ssl_errors()
+        self.verify_ssl = not conf.ignore_ssl_errors()
         if not self.verify_ssl:
             self.logger.debug(u"ATTENTION: Will ignore SSL errors. This might render you vulnerable to attacks.")
             requests.packages.urllib3.disable_warnings()
