+++ sparkmagic/sparkmagic/livyclientlib/sessionmanager.py
@@ -4,12 +4,12 @@ from hdijupyterutils.log import Log
 
 from sparkmagic.livyclientlib.exceptions import SessionManagementException
 from sparkmagic.utils.constants import MAGICS_LOGGER_NAME
-from sparkmagic.utils.configuration import SparkMagicConfiguration
+import sparkmagic.utils.configuration as conf
 
 
 class SessionManager(object):
     def __init__(self):
-        self.logger = Log(MAGICS_LOGGER_NAME, SparkMagicConfiguration().logging_config(), u"SessionManager")
+        self.logger = Log(MAGICS_LOGGER_NAME, conf.logging_config(), u"SessionManager")
 
         self._sessions = dict()
