+++ sparkmagic/sparkmagic/magics/sparkmagicsbase.py
@@ -11,9 +11,9 @@ from IPython.core.magic import Magics, magics_class
 from hdijupyterutils.ipythondisplay import IpythonDisplay
 from hdijupyterutils.log import Log
 
+import sparkmagic.utils.configuration as conf
 from sparkmagic.utils.sparkevents import SparkEvents
 from sparkmagic.utils.constants import MAGICS_LOGGER_NAME
-from sparkmagic.utils.configuration import SparkMagicConfiguration
 from sparkmagic.livyclientlib.sparkcontroller import SparkController
 from sparkmagic.livyclientlib.sqlquery import SQLQuery
 
@@ -24,7 +24,7 @@ class SparkMagicBase(Magics):
         # You must call the parent constructor
         super(SparkMagicBase, self).__init__(shell)
 
-        self.logger = Log(MAGICS_LOGGER_NAME, SparkMagicConfiguration().logging_config(), u"SparkMagics")
+        self.logger = Log(MAGICS_LOGGER_NAME, conf.logging_config(), u"SparkMagics")
         self.ipython_display = IpythonDisplay()
         self.spark_controller = SparkController(self.ipython_display)
