+++ sparkmagic/sparkmagic/tests/test_configuration.py
@@ -2,15 +2,12 @@ from mock import MagicMock
 from nose.tools import assert_equals, assert_not_equals, raises, with_setup
 import json
 
-from sparkmagic.utils.configuration import SparkMagicConfiguration
+import sparkmagic.utils.configuration as conf
 from sparkmagic.livyclientlib.exceptions import BadUserConfigurationException
 
 
 def _setup():
-    global conf
-    
-    conf = SparkMagicConfiguration()
-    conf.override_all(None)
+    conf.override_all({})
     
 
 @with_setup(_setup)
@@ -19,7 +16,7 @@ def test_configuration_override_base64_password():
     overrides = { conf.kernel_python_credentials.__name__: kpc }
     conf.override_all(overrides)
     conf.override(conf.status_sleep_seconds.__name__, 1)
-    assert_equals(conf.overrides, { conf.kernel_python_credentials.__name__: kpc,
+    assert_equals(conf.d, { conf.kernel_python_credentials.__name__: kpc,
                                      conf.status_sleep_seconds.__name__: 1 })
     assert_equals(conf.status_sleep_seconds(), 1)
     assert_equals(conf.base64_kernel_python_credentials(), { 'username': 'U', 'password': 'password', 'url': 'L' })
@@ -31,7 +28,7 @@ def test_configuration_override_fallback_to_password():
     overrides = { conf.kernel_python_credentials.__name__: kpc }
     conf.override_all(overrides)
     conf.override(conf.status_sleep_seconds.__name__, 1)
-    assert_equals(conf.overrides, { conf.kernel_python_credentials.__name__: kpc,
+    assert_equals(conf.d, { conf.kernel_python_credentials.__name__: kpc,
                                      conf.status_sleep_seconds.__name__: 1 })
     assert_equals(conf.status_sleep_seconds(), 1)
     assert_equals(conf.base64_kernel_python_credentials(), kpc)
@@ -43,7 +40,7 @@ def test_configuration_override_work_with_empty_password():
     overrides = { conf.kernel_python_credentials.__name__: kpc }
     conf.override_all(overrides)
     conf.override(conf.status_sleep_seconds.__name__, 1)
-    assert_equals(conf.overrides, { conf.kernel_python_credentials.__name__: kpc,
+    assert_equals(conf.d, { conf.kernel_python_credentials.__name__: kpc,
                                      conf.status_sleep_seconds.__name__: 1 })
     assert_equals(conf.status_sleep_seconds(), 1)
     assert_equals(conf.base64_kernel_python_credentials(),  { 'username': 'U', 'password': '', 'url': '' })
